+++ src/main/java/org/qiyana/model/Customer.java
@@ -1,4 +1,4 @@
-package org.example.model;
+package org.qiyana.model;
 
 import java.util.ArrayList;
 import java.util.List;
